LayoutTests/http/tests/w3c/dom/events/ProgressEvent-expected.txt


FAIL Default event values. assert_equals: expected (boolean) false but got (undefined) undefined
FAIL document.createEvent() should not work with ProgressEvent. assert_throws: function "function () {
    document.createEvent("ProgressEvent")
  }" did not throw
